CR-16: split config_file.rs into directory module
config_file.rs grew to ~1500 lines bundling nine concerns —
schema, loading, unknown-key collection, merging, hot-reload
diff/apply, notification, watching, and print-effective. The
file's own section dividers (// ─── Schema types ───,
// ─── Loading ───, etc.) were the file telling us it wanted to
be split.

Promote to a directory module:

  config_file/
    mod.rs          ConfigError, default_config_path, re-exports
    schema.rs       RawConfigFile + sections + StringOrList
    load.rs         load_config_file + unknown-key collection
    merge.rs        merge + was_set_on_cli
    hot_reload.rs   DiffResult, diff_config, apply_config_change,
                    apply_hot_reloadable_changes, preserve_restart_fields,
                    RESTART_REQUIRED_FIELDS
    notify.rs       notify_user + test-notifier slot
    watch.rs        watch_config_file
    print.rs        print_effective_config

No call sites changed — `crate::config_file::*` paths are stable
via re-exports in mod.rs. Tests travel with their sections
(per CLAUDE.md's `#[cfg(test)] mod tests` at bottom convention).
Cross-sibling visibility uses pub(super); the directory's
public surface stays pub(crate).

Behavior unchanged. All 149 unit + 4 integration tests pass on
the post-PR tip just as they did pre-split.

Refs CR-2026-05-03-16.
